// src/kingdom/types/learning.ts
export interface UnitHierarchyNode {
  id: string;
  title: string;
  description?: string;
  lessons: LessonHierarchyNode[];
}

export interface LevelHierarchyNode {
  id: string;
  title: string;
  description?: string;
  units: UnitHierarchyNode[];
}

export interface SubjectCurriculum {
  id: string;
  subjectName: string;
  iconName: string;
  levels: LevelHierarchyNode[];
}